Define environments 'nonnormative' and 'example', with application to two test cases
In order for this to work with reasonable markup of inline code, one should drop the 'basicstyle' argument to \lstinline, with the consequence that the font size will be the same as for displayed listings.  To me, this looks even better than when normal font size is used, but that's obviously a matter of taste.

The only alternative that I have managed to think of so far would be to change all
  \lstinline[basicstyle=\ttfamily]
to
  \lstinline[basicstyle=\upshape\ttfamily]
but I just find that to be an unacceptable amount of boilerplate code for every inline code snippet.

The environments are not intended for use inside paragraphs, but to stand as paragraphs by themselves.  Hence, even if they could be used inside paragraphs as they are defined now, I would consider it an error to do so, since that would limit how we can improve the styling in the future.