Don't allow language version comment anywhere except at top level
At the 100th design meeting, @HansOlsson wanted us to rethink the earlier decision to allow language version to be specified at different levels of a package hierarchy.  The purpose of this PR is to invite for discussion of this topic, so that we can settle it (again) at the next design meeting.