Drop the top level Figures container in favor of reusing Documentation
Having figures as part of the Documentation has become very natural now that they also include captions, and that part of the current content of the 'info' might end up being migrated to figure captions instead.

Second, the figures are expected to be heavily linked from the 'info', which is another reason for keeping them close together.

Third, this avoids the use of a container record with just a single record member.