Remove confusing whitespace in regular expression
The [poll](https://trac.modelica.org/Modelica/ticket/1726#comment:30) behind the current regular expression was in favor of just having one space after the `//!`.  This was already expressed as `//![ ]`, but was then followed by a space that wasn't inside a character range, effectively mandating (in a very strange way) that there should be two spaces.  The latter space was must have been a mistake in the implementation of the poll, and is now removed.