Forbid hiding loop variables in nested for-equations
Consider the following example model
```
  model Example
    "Example model with nested for loops"
    final parameter Integer M = 3 "problem height";
    final parameter Integer N = 4 "problem width";
    Real[M,N] A "matrix";
    Real[M] x "vector";
    Real[N] y "vector";
  equation
    x = {sin(2*3.1415926*time*i) for i in 1:M};
    y = {cos(2*3.1415926*time*i) for i in 1:N};
    for i in 1:M loop
      for i in 1:N loop
        A[i,i] = x[i]*y[i];
      end for;
    end for;
  end Example;
```
Apparently [Section 8.3.2](https://specification.modelica.org/master/Ch8.html#for-equations-repetitive-equation-structures) of the specification doesn't explicitly forbid this kind of model, so in principle the `i` in the second nested loop could hide the `i` of the first one.  While hiding non-loop variables (which is allowed though discouraged) may make sense, if they are irrelevant in within the scope of the for-equations, hiding loop variables looks like a recipe for disaster, and in most cases will be the result of a plain typing error. The compiler should reject model Example without hesitation.