Move nonnormative content regarding 'revisions' to separate paragraph
This is just the first of many similar changes to come.  This commit mainly serves to populate the cleanup/nonnormative branch with some changes so that a pull request can be initiated.

Note that even this first change fixes a language syntax issue in addition that was missed due to the nonnormative content being in the middle of a paragraph.