Tentatively remove all mathescape=true used with \lstinline
This is just to make the HTML build to work while investigating the problem with mathescape=true.  Obviously, if we don't find a solution, just removing mathescape=true won't be enough; we'll also need change the actual math parts into something else, and then we should take the opportunity to do this consistently.