Use equally sized figures for the different kinds of discrete-time variables
This is a preparation for having these side by side as subfigures, as the non-conforming layout of the old figures didn't look good at all with the graphics appearing so close together.

As a bonus, the new figures also come as SVG, which makes them crisp and scalable also in the LaTeXML build.