Don't exclude directory 'css' from Sublime Text project
With the addition of MLS.css to this directory, there is at least one file that is meant for manual editing.  If there are files in this directory that shouldn't med edited manually, it could be the case that they should be removed from the directory, or that we should have different directories for the ones we should edit manually and the ones we shouldn't.