Add roadmap item for getting rid of 'protected'
The purpose of having `protected` has mostly been lost now when all equations are gathered in the one and only model.  We just need to think about whether another annotation is necessary to indicate that a component was protected in the original full Modelica code, or if `HideResult = true` should be sufficient.