Initial commit of documents for licensing and encryption standardization proposal
This is material from the original proposal somewhat adapted to the new MCP process. it describes how the open source prototype available on https://github.com/modelon-community/SEMLA works. It describes well how it works, but will need some work to be ready as specification text. I can help port this to a different format if needed.