Add special rule for time. Note that it isn't clearly stated as an 'exception', since time is an input and arguable not 'unknown'.
However, 'unknown' is explained as not parameter or constant; and clearly time is neither.
In general it makes sense that inputs should be included in the analysis - but I view that as a separate topic;
as it wouldn't influence how time is treated.

If time was treated as a normal unknown there could only be one partition directly using it, meaning that the goals of having any model in clocked systems would be completely broken.

If time would be ignored for the analysis then use of 'sample(time)' in 16.3 and in Modelica.Clocked.RealSignals.TimeBasedSources.Ramp would be redundant.
That would also remove the 'safety' aspect of normal synchronous models; and wasn't part of the original design.

The new formulation forces time to be in the Continuous-Time or a Clocked Discretized Continuous-Time partition.