In order to support search and copying of ligatures. Closes #3286 It may be that we need to install additional packages on the docker-image (cmap and one more).
It seems we don't use virtual fonts so the more complicated solution isn't needed.
Note that an alternative would be to use mmap, which also adds math-handling.
It has the nice benefit that symbols like \in and \bigcup are automatically handled by copy-paste.
However, since it doesn't handle fractions it seems unclear if it is worth it.