Clarify strange sentence #3729
And also unify noEvent-handling:
I assume that the previous text was a bit odd since it talked about "event-generating" to avoid stating that it only applied outside of noEvent;
I thus found it necessary to add a noEvent-context for the new text.

And unfortunately the noEvent-part had to be unified.